cgroupfs-mount: conflict with procd-ujail
The procd ujail feature expects the unified cgroup hierarchy to be
mounted on /sys/fs/cgroups. Normally, procd takes care of this, but
having cgroupfs-mount installed will cause the unified hierarchy to be
unmounted and mounts the legacy cgroup hierarchy instead. This
completely breaks procd-ujail, so add a conflict to avoid installing
this package when procd-ujail is enabled.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>