stress-ng: add libjpeg dep
Fixes: https://github.com/openwrt/packages/issues/19210

If libjpeg isn't selected by another package, all is fine.
But if it is selected, the stress-ng build will see it and try to build the
jpeg stressor. This would usually fail sometime and link-time.

In any case, it's better to just pick-up libjpeg as a dependency of
stress-ng. If people want to stress their system with this tool, they can
probably expect libjpeg as well.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>